添加注册

+++ src/views/Register.vue
@@ -5,17 +5,18 @@
         margin: 200px auto;
         background-color: #fff;
         width: 350px;
-        height: 300px;
+        height: 400px;
         padding: 20px;
         border-radius: 10px;
       "
     >
       <div style="margin: 20px 0; text-align: center; font-size: 24px">
-        <b>登录</b>
+        <b>注 册</b>
       </div>
       <el-form :model="user" :rules="rules" ref="userForm">
         <el-form-item prop="username">
           <el-input
+            placeholder="请输入用户名"
             size="medium"
             style="margin: 10px 0"
             prefix-icon="el-icon-user"
@@ -24,11 +25,22 @@
         </el-form-item>
         <el-form-item prop="password">
           <el-input
+            placeholder="请输入密码"
             size="medium"
             style="margin: 10px 0"
             prefix-icon="el-icon-lock"
             show-password
             v-model="user.password"
+          ></el-input>
+        </el-form-item>
+        <el-form-item prop="confirmPassword">
+          <el-input
+            placeholder="请确认密码"
+            size="medium"
+            style="margin: 10px 0"
+            prefix-icon="el-icon-lock"
+            show-password
+            v-model="user.confirmPassword"
           ></el-input>
         </el-form-item>
         <el-form-item style="margin: 10px 0; text-align: right">
@@ -37,14 +49,14 @@
             size="small"
             autocomplete="off"
             @click="login"
-            >登录</el-button
+            >注册</el-button
           >
           <el-button
             type="warning"
             size="small"
             autocomplete="off"
-            @click="$router.push('/register')"
-            >注册</el-button
+            @click="$router.push('/login')"
+            >返回登录</el-button
           >
         </el-form-item>
       </el-form>
@@ -54,7 +66,7 @@
 
 <script>
 export default {
-  name: "LoginPage",
+  name: "RegisterPage",
   data() {
     return {
       user: {},
@@ -71,6 +83,15 @@ export default {
         ],
         password: [
           { required: true, message: "请输入密码", trigger: "blur" },
+          {
+            min: 2,
+            max: 20,
+            message: "长度在2到20个字符之间",
+            trigger: "blur",
+          },
+        ],
+        confirmPassword: [
+          { required: true, message: "请确认密码", trigger: "blur" },
           {
             min: 2,
             max: 20,
@@ -107,13 +128,17 @@ export default {
       //   });
       this.$refs["userForm"].validate((valid) => {
         if (valid) {
-          this.request.post("/user/login", this.user).then((res) => {
+          // 表单校验
+          if (this.user.password !== this.user.confirmPassword) {
+            this.$message.error("两次输入的密码不一致");
+            return false;
+          }
+          this.request.post("/user/register", this.user).then((res) => {
             if (res.code === "200") {
               localStorage.setItem("user", JSON.stringify(res.data)); // 将后端传来的用户信息存储到浏览器
-              this.$router.push("/");
-              this.$message.success("登录成功");
+              this.$message.success("注册成功");
             } else {
-              this.$message.error("用户名或密码错误");
+              this.$message.error(res.msg);
             }
           });
         }
